pyproject.toml
[tool.poetry]
name = "educagil"
version = "0.1.0"
description = ""
authors = ["Luiz Fernando <luiz-fernando@outlook.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.2"
djangorestframework = "^3.16.0"
drf-yasg = "^1.21.10"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
pytest-django = "^4.11.1"
ruff = "^0.11.5"
taskipy = "^1.14.1"


[tool.poetry.group.pytest.dependencies]
pytest-cov = "^6.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff] 
line-length = 79
extend-exclude = ['migrations']

[tool.ruff.lint]
preview = true
select = ['I', 'F', 'E', 'W', 'PL', 'PT']

[tool.ruff.format]
preview = true
quote-style = 'single'

[tool.pytest.ini_options]
pythonpath = "."
addopts = '-p no:warnings'

[tool.taskipy.tasks]
run = 'python manage.py runserver'
# pre_test = 'task list'
# test = 'pytest --cov=fast_zero -vv'
# post_test = 'coverage.html'
# lint = 'ruff check . ; ruff check . --diff'
format = 'ruff check . --fix ; ruff format .'
